Fix tests and move
